remove unused hooks and migrate interfaces to hookSystem
